rename appendIdentityToUrl to redirectWithIdentity
