+++ src/lib/actions/redirectWithIdentity/index.js
@@ -10,9 +10,9 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const createAppendIdentityToUrl = require("./createAppendIdentityToUrl");
+const createRedirectWithIdentity = require("./createRedirectWithIdentity");
 const instanceManager = require("../../instanceManager/index");
 
 const document = window.document;
 
-module.exports = createAppendIdentityToUrl({ instanceManager, document });
+module.exports = createRedirectWithIdentity({ instanceManager, document });
