+++ src/view/actions/redirectWithIdentity.html
@@ -8,6 +8,6 @@
   <body>
     <div id="root"></div>
     <script src="https://assets.adobedtm.com/activation/reactor/extensionbridge/extensionbridge.min.js"></script>
-    <script type="module" src="./appendIdentityToUrl.jsx"></script>
+    <script type="module" src="./redirectWithIdentity.jsx"></script>
   </body>
 </html>
